rename again to handlers as in garmin
